feat: add missing routes
