+++ src/pages/NewProgram/Creator/Creator.tsx
@@ -1,0 +1,8 @@
+import React, { memo } from "react";
+
+const CreatorPage: React.FC = () => {
+  return <div>Creator</div>;
+};
+
+const Creator = memo(CreatorPage);
+export default Creator;
